backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3
python-jose==3.3.0
python-multipart==0.0.6
passlib==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.21
psycopg2-binary==2.9.7
motor==3.3.1
pymongo==4.5.0
emails==0.6
jinja2==3.1.2
aiohttp==3.8.5
google-generativeai==0.3.1 